music/search: Fix failed interaction after choose song
When user choose song from search result, the interaction of search
result will be deleted but it failed to send "Now Playing" interaction
(for this case, the song is the first play song) because bot send
defered response as interaction response. It makes discord will shown
error because interaction was replied.

Remove defered interaction response, so that response will be replied by
onPlay handler or queue handler.

Signed-off-by: Taufiq Pohan <moonchild@gnuweeb.org>